+++ src/main/java/com/epam/ta/reportportal/core/item/merge/MergeTestItemHandlerImpl.java
@@ -41,15 +41,14 @@ import com.epam.ta.reportportal.ws.model.item.MergeTestItemRQ;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
+import java.util.stream.Collectors;
 
 import static com.epam.ta.reportportal.commons.Predicates.equalTo;
 import static com.epam.ta.reportportal.commons.Predicates.notNull;
 import static com.epam.ta.reportportal.commons.validation.BusinessRule.expect;
 import static com.epam.ta.reportportal.ws.model.ErrorType.*;
+
 
 @Service
 public class MergeTestItemHandlerImpl implements MergeTestItemHandler {
@@ -89,22 +88,51 @@ public class MergeTestItemHandlerImpl implements MergeTestItemHandler {
             validateTestItemInProject(itemToMerge, project);
             itemsToMerge.add(itemToMerge);
         }
+
         MergeStrategyType mergeStrategyType = MergeStrategyType.fromValue(rq.getMergeStrategyType());
         expect(mergeStrategyType, Predicates.notNull()).verify(ErrorType.UNSUPPORTED_MERGE_STRATEGY_TYPE, rq.getMergeStrategyType());
-
         MergeStrategy mergeStrategy = mergeStrategyFactory.getStrategy(mergeStrategyType);
         mergeStrategy.mergeTestItems(testItemTarget, itemsToMerge);
 
-        StatisticsFacade statisticsFacade = statisticsFacadeFactory.getStatisticsFacade(project.getConfiguration().getStatisticsCalculationStrategy());
+        updateTargetItemInfo(testItemTarget, itemsToMerge);
 
+        StatisticsFacade statisticsFacade = statisticsFacadeFactory
+                .getStatisticsFacade(project.getConfiguration().getStatisticsCalculationStrategy());
         for (String launchID : sourceLaunches) {
             Launch launch = launchRepository.findOne(launchID);
             statisticsFacade.recalculateStatistics(launch);
         }
-
         statisticsFacade.recalculateStatistics(launchTarget);
 
         return new OperationCompletionRS("TestItem with ID = '" + item + "' successfully merged.");
+    }
+
+    /**
+     * Collects tags and descriptions from items and add them to target. Same tags
+     * and descriptions are added only once. Updates start and end times of target.
+     * @param target item to be merged
+     * @param items items to merge
+     */
+    private void updateTargetItemInfo(TestItem target, List<TestItem> items) {
+        items.stream().map(it -> Optional.ofNullable(it.getTags())).reduce((reduced, actual) -> {
+            reduced.orElse(new HashSet<>()).addAll(actual.get());
+            return reduced;
+        }).ifPresent(it
+                -> Optional.ofNullable(target.getTags()).orElse(new HashSet<>())
+                .addAll(it.orElse(Collections.emptySet())));
+
+        StringBuilder result = new StringBuilder(target.getItemDescription());
+        String collect = items.stream().map(TestItem::getItemDescription).filter(description
+                -> !description.equals(target.getItemDescription())).collect(Collectors.joining("\n"));
+        if (!collect.isEmpty()) {
+            target.setItemDescription(result.append("\n").append(collect).toString());
+        }
+
+        items.add(target);
+        items.sort(Comparator.comparing(TestItem::getStartTime));
+        target.setStartTime(items.get(0).getStartTime());
+        target.setEndTime(items.get(items.size()-1).getEndTime());
+        testItemRepository.save(target);
     }
 
     private void validateLaunchInProject(Launch launch, Project project) {